edit to 101 -3

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -21,7 +21,7 @@ void print_number(int n)
 
 void print_num(int n)
 {
-	int divisor;
+	long int divisor;
 
 	if (n == 0)
 	{
